backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
python-dotenv==1.0.1
httpx==0.27.2
anthropic==0.40.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.151.0
Pillow==11.0.0
aiofiles==24.1.0
